refactor(browserslist-config-growcss): changed name
