+++ packages/coding-standard/browserslist-config-growcss/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@growcss/browserslist-config",
+  "name": "browserslist-config-growcss",
   "version": "1.0.0",
   "description": "Shared GrowCss browserlist config.",
   "keywords": [
